+++ .gitignore
@@ -47,3 +47,5 @@
 **/package-lock.json
 
 **/node_modules/
+
+**/dist/
