+++ registration/src/index.html
@@ -29,8 +29,5 @@
       <button type="submit">Update Password</button>
     </form>
   </div>
-
-  <!-- Use "type=module" to support ES6 imports -->
-  <script type="module" src="app.js"></script>
 </body>
 </html>
