+++ registration/src/style.css
